[Silver III] Title: 주유소, Time: 360 ms, Memory: 41472 KB, Score: 100 point -BaekjoonHub

+++ 백준/Silver/13305. 주유소/주유소.java
@@ -7,19 +7,19 @@ public class Main {
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         int N = Integer.parseInt(br.readLine());
-        int[] distance = new int[N - 1];
+        long[] distance = new long[N - 1];
         StringTokenizer st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N - 1; i++) {
-            distance[i] = Integer.parseInt(st.nextToken());
+            distance[i] = Long.parseLong(st.nextToken());
         }
-        int[] price = new int[N];
+        long[] price = new long[N];
         st = new StringTokenizer(br.readLine());
         for (int i = 0; i < N; i++) {
-            price[i] = Integer.parseInt(st.nextToken());
+            price[i] = Long.parseLong(st.nextToken());
         }
 
-        int sum = 0;
-        int prevPrice = Integer.MAX_VALUE;
+        long sum = 0;
+        long prevPrice = Long.MAX_VALUE;
         for (int i = 0; i < price.length - 1; i++) {
             if (price[i] < prevPrice) {
                 prevPrice = price[i];
